DOC/BUILD: Improve docs and doc build process.

+++ tools/sphinxify_notebooks.py
@@ -17,6 +17,8 @@ from __future__ import division, print_function, absolute_import
 import os
 import sys
 import shutil
+import subprocess
+import re
 
 from IPython.nbconvert import RSTExporter
 from IPython.nbconvert.writers import FilesWriter
@@ -32,70 +34,28 @@ try:
 except NameError:
     WindowsError = None
 
+def convert_notebook(fromfile, tofile, output_subdir):
+    # ipython nbconvert --to rst --execute fromfile --output tofile
+    r = subprocess.call(['ipython nbconvert --to rst --execute %s --output %s' %(fromfile, tofile)], shell=True)
+    if r != 0:
+        raise Exception('Sphinxify failed')
 
-def _prepare_execute(file, _nbformat=None):
-    if _nbformat is None:
-        _nbformat = nbformat.current
-    return _nbformat.reads_json(open(file).read())
+    # Replace absolute (image) file paths with relative paths
+    with open(tofile + '.rst', 'r') as f:
+        string = f.read()
+        string = re.sub(output_subdir, '', string)
 
-
-def _execute_runipy(file):
-    notebook = _prepare_execute(file)
-
-    from runipy.notebook_runner import NotebookRunner
-
-    runner = NotebookRunner(notebook)
-    runner.run_notebook(skip_exceptions=True)
-
-    return runner.nb, {}
-
-
-def _execute_ipython(file, resources=None):
-    notebook = _prepare_execute(file)
-
-    from IPython.nbconvert.preprocessors import ExecutePreprocessor
-    if resources is None:
-        resources = {}
-    notebook, resources = ExecutePreprocessor().preprocess(notebook, resources)
-
-    return notebook, resources
-
-
-def convert_notebook(notebook, tofile, resources=None):
-    todir = os.path.dirname(tofile)
-
-    exporter = RSTExporter()
-    if resources is None:
-        resources = {}
-    resources.setdefault('unique_key', os.path.basename(tofile).replace('.', '_'))
-    output, resources = exporter.from_notebook_node(notebook,
-                                                    resources=resources)
-
-    writer = FilesWriter()
-    writer.build_directory = todir
-    writer.write(output, resources, notebook_name=tofile)
-
+    with open(tofile + '.rst', 'w') as f:
+        f.write(string)
 
 def find_process_files(source_dir, output_dir):
     # Figure out what processors we have available
     try:
-        import runipy
-        execute_runipy = True
-        execute_notebook = _execute_runipy
-    except ImportError:
-        execute_runipy = False
-    try:
         import IPython
-        execute_ipython = int(IPython.__version__.split('.')[0]) >= 3
-        if execute_ipython:
-            execute_notebook = _execute_ipython
+        if not int(IPython.__version__.split('.')[0]) >= 3:
+            raise ImportError
     except ImportError:
-        raise Exception('IPython required to build documentation.')
-
-    # Make sure one of them is available
-    if not execute_ipython and not execute_runipy:
-        raise Exception('Either IPython >= 3.0.0 or runipy required to'
-                        ' build documentation.')
+        raise Exception('IPython >= 3.0.0 required to build documentation.')
 
     # Walk through the notebooks
     this_dir = os.path.abspath(os.getcwd())
@@ -103,6 +63,8 @@ def find_process_files(source_dir, output_dir):
         if cur_dir.endswith(os.sep + '.ipynb_checkpoints'):
             continue
         rel_cur_dir = os.path.relpath(cur_dir, source_dir)
+        if rel_cur_dir == '.':
+            rel_cur_dir = ''
         for filename in files:
             if filename.endswith('.ipynb'):
                 # Need to change the working directory to the one in which
@@ -120,8 +82,7 @@ def find_process_files(source_dir, output_dir):
                 tofile = os.path.join(output_subdir, filename)
 
                 # Run and convert the notebook
-                notebook, resources = execute_notebook(fromfile)
-                convert_notebook(notebook, tofile, resources)
+                convert_notebook(fromfile, tofile, output_subdir)
         os.chdir(this_dir)
 
 
@@ -130,11 +91,11 @@ def main():
     try:
         source_dir = sys.argv[1]
     except IndexError:
-        source_dir = DEFAULT_SOURCE
+        source_dir = os.path.abspath(DEFAULT_SOURCE)
     try:
         output_loc = sys.argv[2]
     except IndexError:
-        output_loc = DEFAULT_OUTPUT
+        output_loc = os.path.abspath(DEFAULT_OUTPUT)
 
     # Double check source directory actually exists
     if not os.path.exists(source_dir):
@@ -142,11 +103,13 @@ def main():
 
     # Clean the output directory
     output_dir = os.path.join(output_loc, DEFAULT_GEN_DIR)
-    try:
-        shutil.rmtree(output_dir)
-    except:
-        pass
-    os.makedirs(output_dir)
+    delete = raw_input('Are you sure you want to delete %s? [y/n]: ' % output_dir).lower()
+    if delete in ['y', 'yes']:
+        try:
+            shutil.rmtree(output_dir)
+        except Exception as e:
+            pass
+        os.makedirs(output_dir)
 
     # Get absolute paths
     source_dir = os.path.abspath(source_dir)
